refactor: Extract operators map
