+++ src/components/map.stories.tsx
@@ -1,272 +1,17 @@
-import { GeoJsonLayer } from "@deck.gl/layers/typed";
-import DeckGL, { DeckGLRef } from "@deck.gl/react/typed";
 import { Box, List, ListItemButton } from "@mui/material";
 import { Decorator } from "@storybook/react";
-import * as turf from "@turf/turf";
-import { easeExpIn, mean, median, ScaleThreshold } from "d3";
-import { Feature, Geometry, MultiPolygon, Polygon } from "geojson";
-import { keyBy } from "lodash";
-import { useCallback, useMemo, useRef, useState } from "react";
+import { median } from "d3";
+import { useMemo, useState } from "react";
 import { createClient, Provider } from "urql";
 
 import { ChoroplethMap } from "src/components/map";
 import { MapProvider } from "src/components/map-context";
-import { getFillColor, getZoomFromBounds } from "src/components/map-helpers";
-import {
-  getOperatorsFeatureCollection,
-  MunicipalityFeatureCollection,
-  OperatorLayerProperties,
-  useGeoData,
-} from "src/data/geo";
-import { useFetch } from "src/data/use-fetch";
-import { ElectricityCategory, useColorScale } from "src/domain/data";
+import OperatorsMap, {
+  DisplayedAttribute,
+  displayedAttributes,
+} from "src/components/operators-map";
+import { useColorScale } from "src/domain/data";
 import { useSunshineTariffQuery } from "src/graphql/queries";
-import { SunshineDataRow } from "src/graphql/resolver-types";
-import { truthy } from "src/lib/truthy";
-import { getOperatorsMunicipalities } from "src/rdf/queries";
-
-import { props } from "./map.mock";
-
-const TRANSPARENT = [255, 255, 255, 0] as [number, number, number, number];
-
-const colorAccessor = (d: { value: number }) => d.value;
-
-export const Example = () => {
-  const colorScale = useColorScale({
-    observations: props.observations,
-    medianValue: props.medianValue,
-    accessor: colorAccessor,
-  });
-  return (
-    <Box
-      width="800px"
-      height="800px"
-      position="relative"
-      sx={{
-        "& #deckgl-wrapper": {
-          width: "100%",
-          height: "100%",
-        },
-      }}
-    >
-      <ChoroplethMap
-        {...props}
-        colorScale={colorScale}
-        onMunicipalityLayerClick={() => {}}
-      />
-    </Box>
-  );
-};
-
-const sunshineAttributeToElectricityCategory: Partial<
-  Record<keyof SunshineDataRow, ElectricityCategory>
-> = {
-  tariffEC2: "C2",
-  tariffEC3: "C3",
-  tariffEC4: "C4",
-  tariffEC6: "C6",
-  tariffEH2: "H2",
-  tariffEH4: "H4",
-  tariffEH7: "H7",
-  tariffNC2: "C2",
-  tariffNC3: "C3",
-  tariffNC4: "C4",
-  tariffNC6: "C6",
-  tariffNH2: "H2",
-  tariffNH4: "H4",
-  tariffNH7: "H7",
-};
-
-const displayedAttributes = [
-  "tariffEC2",
-  "tariffEC3",
-  "tariffEC4",
-  "tariffEC6",
-  "tariffEH2",
-  "tariffEH4",
-  "tariffEH7",
-  "tariffNC2",
-  "tariffNC3",
-  "tariffNC4",
-  "tariffNC6",
-  "tariffNH2",
-  "tariffNH4",
-  "tariffNH7",
-  "saidiTotal",
-  "saidiUnplanned",
-  "saifiTotal",
-  "saifiUnplanned",
-] satisfies (keyof SunshineDataRow)[];
-
-type DisplayedAttribute = (typeof displayedAttributes)[number];
-
-type OperatorsMapProps = {
-  period: string;
-  colorScale: ScaleThreshold<number, string, never>;
-  attribute: DisplayedAttribute;
-  observations?: SunshineDataRow[];
-};
-
-const OperatorsMap = ({
-  period,
-  colorScale,
-  attribute,
-  observations,
-}: OperatorsMapProps) => {
-  const deckglRef = useRef<DeckGLRef>(null);
-  const electricityCategory =
-    sunshineAttributeToElectricityCategory[attribute] ?? "all";
-  const { data: operatorMunicipalities } = useFetch({
-    key: `operator-municipalities-${period}-${electricityCategory}`,
-    queryFn: () => getOperatorsMunicipalities(period, electricityCategory),
-  });
-  const { data: geoData } = useGeoData(period);
-
-  const observationsByOperator = useMemo(() => {
-    return keyBy(observations ?? [], "operatorId");
-  }, [observations]);
-
-  console.log({ observationsByOperator, attribute });
-
-  const enhancedGeoData = useMemo(() => {
-    if (!operatorMunicipalities || !geoData) {
-      return null;
-    }
-    return getOperatorsFeatureCollection(
-      operatorMunicipalities,
-      geoData?.municipalities as MunicipalityFeatureCollection
-    );
-  }, [operatorMunicipalities, geoData]);
-
-  const getMapFillColor = useCallback(
-    (x: Feature<Geometry, OperatorLayerProperties>) => {
-      if (!x.properties) {
-        return TRANSPARENT;
-      }
-      const operatorIds = x.properties.operators;
-      const values = operatorIds
-        .map((x) => {
-          const op = observationsByOperator[x];
-          return op?.[attribute] ?? null;
-        })
-        .filter(truthy);
-      if (values.length === 0) {
-        return TRANSPARENT;
-      }
-      const value = mean(values);
-      console.log(
-        `getMapFillColor: attribute=${attribute}, value=${value}, operatorIds=${operatorIds}`
-      );
-      const color = getFillColor(colorScale, value, false);
-      console.log(color);
-      return color;
-    },
-    [attribute, colorScale, observationsByOperator]
-  );
-
-  if (!geoData || !geoData.municipalities || !enhancedGeoData) {
-    return null;
-  }
-
-  return (
-    <DeckGL
-      initialViewState={{
-        latitude: 46.8182,
-        longitude: 8.2275,
-        zoom: 7,
-        maxZoom: 16,
-        minZoom: 2,
-        pitch: 0,
-        bearing: 0,
-      }}
-      getTooltip={({ object }) => {
-        if (!object) {
-          return null;
-        }
-        const operatorIds = (object.properties as OperatorLayerProperties)
-          ?.operators;
-        return {
-          html: `<div>${operatorIds.join("<br/>")}</div>`,
-          style: {
-            backgroundColor: "white",
-            color: "black",
-            fontSize: "12px",
-            padding: "5px",
-          },
-        };
-      }}
-      controller={true}
-      ref={deckglRef}
-      layers={[
-        new GeoJsonLayer<
-          Feature<Polygon | MultiPolygon, OperatorLayerProperties>
-        >({
-          id: "operator-layer",
-          data: enhancedGeoData.features,
-          filled: true,
-          autoHighlight: true,
-          highlightColor: [0, 0, 0, 100],
-          updateTriggers: {
-            getFillColor: [getMapFillColor],
-          },
-          getFillColor: (x) => getMapFillColor(x),
-          getLineColor: [255, 255, 255, 100],
-          getLineWidth: 1.5,
-          lineWidthUnits: "pixels",
-          onClick: (info) => {
-            if (info.object && info.viewport) {
-              const bounds = turf.bbox(info.object.geometry);
-              deckglRef.current?.deck?.setProps({
-                viewState: {
-                  longitude: (bounds[0] + bounds[2]) / 2,
-                  latitude: (bounds[1] + bounds[3]) / 2,
-                  zoom: getZoomFromBounds(bounds, info.viewport),
-                  transitionDuration: 300,
-                },
-              });
-            }
-          },
-          transitions: {
-            getFillColor: {
-              duration: 300,
-              easing: easeExpIn,
-            },
-          },
-        }),
-
-        // Transparent layer only used for hover effect
-        new GeoJsonLayer<
-          Feature<Polygon | MultiPolygon, OperatorLayerProperties>
-        >({
-          id: "operator-layer-pickable",
-          data: enhancedGeoData.features,
-          filled: true,
-          autoHighlight: true,
-          stroked: false,
-          highlightColor: [0, 0, 0, 100],
-          updateTriggers: {
-            getFillColor: [getMapFillColor],
-          },
-          getFillColor: TRANSPARENT,
-          lineWidthUnits: "pixels",
-          pickable: true,
-        }),
-
-        // Municipality Layer
-        new GeoJsonLayer({
-          id: "municipality-layer",
-          data: geoData.municipalities.features,
-          getLineColor: [255, 255, 255],
-          highlightColor: [0, 0, 255],
-          lineWidthUnits: "pixels",
-          stroked: true,
-          autoHighlight: true,
-          getLineWidth: 0.25,
-        }),
-      ]}
-    />
-  );
-};
 
 export const Operators = () => {
   const period = "2024";
